requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0
rapidfuzz>=3.0.0
xlrd>=2.0.1
PySide6>=6.6.0
